:recycle: Move updateTransactionsIfNeeded back to account actions
